+++ src/resolvers/affiliation.ts
@@ -12,8 +12,7 @@ import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError
 import { prepareObjectForLogs } from "../logger";
 import { GraphQLError } from "graphql";
 import { PaginationOptionsForCursors, PaginationOptionsForOffsets, PaginationType } from "../types/general";
-import { isNullOrUndefined } from "../utils/helpers";
-import {formatISO9075} from "date-fns";
+import { isNullOrUndefined, normaliseDateTime } from "../utils/helpers";
 
 export const resolvers: Resolvers = {
   Query: {
@@ -158,10 +157,10 @@ export const resolvers: Resolvers = {
 
   Affiliation: {
     created: (parent: Affiliation) => {
-      return formatISO9075(new Date(parent.created));
+      return normaliseDateTime(parent.created);
     },
     modified: (parent: Affiliation) => {
-      return formatISO9075(new Date(parent.modified));
+      return normaliseDateTime(parent.modified);
     }
   }
 }