+++ src/resolvers/output.ts
@@ -11,7 +11,7 @@ import { AuthenticationError, ForbiddenError, InternalServerError, NotFoundError
 import { hasPermissionOnProject } from '../services/projectService';
 import { OutputType } from '../models/OutputType';
 import { GraphQLError } from 'graphql';
-import {formatISO9075} from "date-fns";
+import { normaliseDateTime } from "../utils/helpers";
 
 // Process updates to the Repository associations
 async function processRepositoryUpdates(
@@ -346,10 +346,10 @@ export const resolvers: Resolvers = {
       );
     },
     created: (parent: ProjectOutput) => {
-      return formatISO9075(new Date(parent.created));
+      return normaliseDateTime(parent.created);
     },
     modified: (parent: ProjectOutput) => {
-      return formatISO9075(new Date(parent.modified));
+      return normaliseDateTime(parent.modified);
     }
   },
 };