+++ src/resolvers/versionedQuestionCondition.ts
@@ -8,8 +8,7 @@ import { AuthenticationError, ForbiddenError, InternalServerError } from "../uti
 import { prepareObjectForLogs } from "../logger";
 import { isAdmin } from "../services/authService";
 import { GraphQLError } from "graphql";
-import {formatISO9075} from "date-fns";
-
+import { normaliseDateTime } from "../utils/helpers";
 
 export const resolvers: Resolvers = {
   Query: {
@@ -38,10 +37,10 @@ export const resolvers: Resolvers = {
 
   VersionedQuestionCondition: {
     created: (parent: VersionedQuestionCondition) => {
-      return formatISO9075(new Date(parent.created));
+      return normaliseDateTime(parent.created);
     },
     modified: (parent: VersionedQuestionCondition) => {
-      return formatISO9075(new Date(parent.modified));
+      return normaliseDateTime(parent.modified);
     }
   },
 };